+++ griptape/tasks/code_execution_task.py
@@ -14,5 +14,5 @@ if TYPE_CHECKING:
 class CodeExecutionTask(BaseTextInputTask):
     run_fn: Callable[[CodeExecutionTask], BaseArtifact] = field(kw_only=True)
 
-    def run(self) -> BaseArtifact:
+    def try_run(self) -> BaseArtifact:
         return self.run_fn(self)