+++ tests/mocks/mock_audio_input_task.py
@@ -6,5 +6,5 @@ from griptape.tasks.base_audio_input_task import BaseAudioInputTask
 
 @define
 class MockAudioInputTask(BaseAudioInputTask):
-    def run(self) -> TextArtifact:
+    def try_run(self) -> TextArtifact:
         return TextArtifact(self.input.to_text())