+++ tests/unit/tasks/test_actions_subtask.py
@@ -220,7 +220,7 @@ class TestActionsSubtask:
         task = ToolkitTask(tools=[MockTool()])
         Agent().add_task(task)
         subtask = task.add_subtask(ActionsSubtask(valid_input))
-        subtask.execute()
+        subtask.run()
 
         assert isinstance(subtask.output, ListArtifact)
         assert isinstance(subtask.output.value[0], TextArtifact)
@@ -235,7 +235,7 @@ class TestActionsSubtask:
         task = ToolkitTask(tools=[MockTool()])
         Agent().add_task(task)
         subtask = task.add_subtask(ActionsSubtask(valid_input))
-        subtask.execute()
+        subtask.run()
 
         assert isinstance(subtask.output, ListArtifact)
         assert isinstance(subtask.output.value[0], ErrorArtifact)